+++ modules/apps/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/internal/exportimport/data/handler/test/util/FileEntryRemoteStagingTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.document.library.internal.exportimport.data.handler.test;
+package com.liferay.document.library.internal.exportimport.data.handler.test.util;
 
 import com.liferay.document.library.constants.DLPortletKeys;
 import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
