+++ modules/apps/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/service/test/util/DLFileEntryServiceTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.document.library.service.test;
+package com.liferay.document.library.service.test.util;
 
 import com.liferay.portal.configuration.test.util.ConfigurationTemporarySwapper;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
